// types/character-classes.ts
export enum CharacterClassesEnum {
	Barbarian = 'barbarian',
	Cleric = 'cleric',
	Bard = 'bard',
	Druid = 'druid',
	Monk = 'monk',
	Fighter = 'fighter',
	Paladin = 'paladin',
	Ranger = 'ranger',
	Rogue = 'rogue',
	Sorcerer = 'sorcerer',
	Warlock = 'warlock',
	Wizard = 'wizard',
}

export const characterClasses = [
	{ index: CharacterClassesEnum.Barbarian, name: 'Barbarian' },
	{ index: CharacterClassesEnum.Cleric, name: 'Cleric' },
	{ index: CharacterClassesEnum.Bard, name: 'Bard' },
	{ index: CharacterClassesEnum.Druid, name: 'Druid' },
	{ index: CharacterClassesEnum.Monk, name: 'Monk' },
	{ index: CharacterClassesEnum.Fighter, name: 'Fighter' },
	{ index: CharacterClassesEnum.Paladin, name: 'Paladin' },
	{ index: CharacterClassesEnum.Ranger, name: 'Ranger' },
	{ index: CharacterClassesEnum.Rogue, name: 'Rogue' },
	{ index: CharacterClassesEnum.Sorcerer, name: 'Sorcerer' },
	{ index: CharacterClassesEnum.Warlock, name: 'Warlock' },
	{ index: CharacterClassesEnum.Wizard, name: 'Wizard' },
];